feat: 장소 리스트 호출 메서드 구현 완료
