+++ LeaveGo/Shared/Model/NetworkError.swift
@@ -1,0 +1,14 @@
+//
+//  NetworkError.swift
+//  LeaveGo
+//
+//  Created by Kitcat Seo on 6/10/25.
+//
+
+import Foundation
+
+enum NetworkError: Error {
+    case invalidURL
+    case invalidResponse
+    case decodingError
+}
